pyproject.toml
[project]
name = "unova"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "connectorx>=0.4.2",
    "pandas>=2.2.3",
    "pillow>=11.2.0",
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "sqlalchemy>=2.0.40",
]
